:recycle: "Refactored all api types and decoders into their own seperate file"
Det ble litt uryddig når alle api modulene implementerte og exporterte sine egne typer og decoders. Samlet alle api type definisjoner i types.ts og alle type decoders i decoders.ts
