+++ src/lib/api/registration.ts
@@ -1,90 +1,9 @@
 import axios from 'axios';
-import { array, boolean, decodeType, number, optional, Pojo, record, string } from 'typescript-json-decoder';
+import { array } from 'typescript-json-decoder';
 import handleError from './errors';
+import { responseDecoder, registrationDecoder, spotRangeCountDecoder } from './decoders';
+import { Degree, Answer, Response, Registration, SpotRangeCount } from './types';
 import { HappeningType } from '.';
-
-enum Degree {
-    DTEK = 'DTEK',
-    DSIK = 'DSIK',
-    DVIT = 'DVIT',
-    BINF = 'BINF',
-    IMO = 'IMO',
-    IKT = 'IKT',
-    KOGNI = 'KOGNI',
-    INF = 'INF',
-    PROG = 'PROG',
-    ARMNINF = 'ARMNINF',
-    POST = 'POST',
-    MISC = 'MISC',
-}
-
-const degreeDecoder = (value: Pojo): Degree => {
-    const str: string = string(value);
-
-    switch (str) {
-        case 'DTEK':
-            return Degree.DTEK;
-        case 'DSIK':
-            return Degree.DSIK;
-        case 'DVIT':
-            return Degree.DVIT;
-        case 'BINF':
-            return Degree.BINF;
-        case 'IMO':
-            return Degree.IMO;
-        case 'IKT':
-            return Degree.IKT;
-        case 'KOGNI':
-            return Degree.KOGNI;
-        case 'INF':
-            return Degree.INF;
-        case 'PROG':
-            return Degree.PROG;
-        case 'ARMINF':
-            return Degree.ARMNINF;
-        case 'POST':
-            return Degree.POST;
-        default:
-            return Degree.MISC;
-    }
-};
-
-type Answer = decodeType<typeof answerDecoder>;
-const answerDecoder = record({
-    question: string,
-    answer: string,
-});
-
-type Registration = decodeType<typeof registrationDecoder>;
-const registrationDecoder = record({
-    email: string,
-    firstName: string,
-    lastName: string,
-    degree: degreeDecoder,
-    degreeYear: number,
-    slug: string,
-    terms: boolean,
-    submitDate: string,
-    waitList: boolean,
-    answers: array(answerDecoder),
-});
-
-type Response = decodeType<typeof responseDecoder>;
-const responseDecoder = record({
-    code: string,
-    title: string,
-    desc: string,
-    date: optional(string),
-});
-
-type SpotRangeCount = decodeType<typeof spotRangeCountDecoder>;
-const spotRangeCountDecoder = record({
-    spots: number,
-    minDegreeYear: number,
-    maxDegreeYear: number,
-    regCount: number,
-    waitListCount: number,
-});
 
 const genericError: { title: string; desc: string; date: string | undefined } = {
     title: 'Det har skjedd en feil.',
@@ -222,5 +141,5 @@ const RegistrationAPI = {
     },
 };
 
-export { Degree, RegistrationAPI, registrationRoute };
-export type { Answer, FormValues, Registration, Response, SpotRangeCount };
+export { RegistrationAPI, registrationRoute };
+export type { FormValues };
